requirements.txt
awscli==1.37.6
boto3==1.36.6
botocore==1.36.6
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
docutils==0.16
Events==0.5
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
opensearch-py==2.8.0
packaging==24.2
pluggy==1.5.0
psycopg==3.2.4
pyasn1==0.6.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.11.2
six==1.17.0
urllib3==2.3.0
sqlparse==0.5.3
